Move lib/exceptions to lib/mastodon
